requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Filtering
django-filter>=23.5,<24.0

# Environment variables
python-dotenv>=1.0,<2.0

# Development and Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code Quality
black>=24.0,<25.0
flake8>=7.0,<8.0
isort>=5.13,<6.0
